% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diff.R
\name{calculate_dist}
\alias{calculate_dist}
\title{Compute the distance between two covariance matrices}
\usage{
calculate_dist(covA, covB)
}
\arguments{
\item{covA}{A positive definite, real, square, symmetric matrix.}

\item{covB}{A positive definite, real, square, symmetric matrix.}
}
\value{
The distance between clustA and clustB, based on the metric defined
by Förstner & Moonen, 2003.
}
\description{
Compute the distance between two covariance matrices as described by Förstner
& Moonen, 2003. In particular, the squared distance
\eqn{d^{2}\left(\textbf{A, B}\right)} between two symmetric positive
matrices \eqn{\textbf{A}} and \eqn{\textbf{B}} are defined as follows:
\deqn{\text{tr}\left(\ln^{2}\left(\sqrt{\textbf{A}^{-1}}\textbf{B}
 \sqrt{\textbf{A}^{-1}}\right)\right)}.
In layman's terms, this equation computes how much 'stretching' would be
needed to transform matrix \eqn{B} to \eqn{A}, and sums this stretching up
for all directions. This sum is then the distance between the two matrices.
Though intended to be applied to covariance matrices, this function
would work for any two positive, real, square, symmetric (definite)
matrices.
}
\examples{
# initialise two positive definite symmetric square matrices
matA <- matrix(data = c(2, 1, 1, 2), nrow = 2, ncol = 2)
matB <- matrix(data = c(3, 2, 2, 3), nrow = 2, ncol = 2)
matC <- matrix(data = c(2, 1, 1, 2), nrow = 2, ncol = 2) # same as A

# compute the distance between the two matrices
calculate_dist(matA, matB)

# this should return 0
calculate_dist(matA, matC)

}
\references{
Förstner, W., & Moonen, B. (2003). A Metric for Covariance Matrices.
In Geodesy-The Challenge of the 3rd Millennium (pp. 299–309).
Springer Berlin Heidelberg. https://doi.org/10.1007/978-3-662-05296-9_31
}
