% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{checkEverything}
\alias{checkEverything}
\title{Check the data for zero-inflation or missing values}
\usage{
checkEverything(
  data,
  check_zero = TRUE,
  overall_zero_threshold = 0.5,
  cell_zero_threshold = 0.5,
  gene_zero_threshold = 0.5,
  clean_zero = TRUE,
  clean_NA = TRUE
)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{check_zero}{A logical indicating whether to check for zero-inflation.
Default is TRUE.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}

\item{clean_zero}{A logical indicating whether to remove zero-inflated genes.
and cells. Default is TRUE.}

\item{clean_NA}{A logical indicating whether to impute missing values.
Default is TRUE. This can be changed to FALSE if the user wishes to handle
missing values on their own.}
}
\value{
\code{NULL}.
}
\description{
Checks data for any NA or NaN values, and 'cleans' them by imputation based
on mean gene expression, unless specified to not clean them. If missing
data is found but not cleaned, function will stop execution and raise an
error, because missing values will interfere with analyses. Also, if
specified, can check data for possible zero-inflation by comparing
the number of zero values in the data to a threshold (default=0.5), for which
the data is considered zero-inflated if the number of zero values observed
in the data exceeds the expected number as computed from the threshold.
Zero-inflated genes and cells will then be removed unless specified not to.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib

# change some of the data to zero
df_zero_inflated <- data.frame(df)
df_zero_inflated[,2:4] <- 0
checkEverything(df_zero_inflated)
head(df_zero_inflated)

# change one of the values to NA
df_one_NA <- data.frame(df)
df_one_NA[1,2] <- NA
checkEverything(df_one_NA)
df_one_NA[1,2]

#change one of the values to NaN
df_one_NaN <- data.frame(df)
df_one_NaN[1,2] <- 1/0
checkEverything(df_one_NaN)
df_one_NaN[1,2]

}
