% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{checkMissingData}
\alias{checkMissingData}
\title{Check for NA or NaN values}
\usage{
checkMissingData(data)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}
}
\value{
A boolean value representing whether they are any NA or NaN values
in data
}
\description{
Check for NA or NaN values
}
\examples{
add(1, 1)
add(10, 1)
}
