% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{checkZeroInflation}
\alias{checkZeroInflation}
\title{Check for possible zero-inflation in the data}
\usage{
checkZeroInflation(data, overall_zero_threshold = 0.5)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}
}
\value{
A boolean value representing whether the proportion of zeroes in the
data is greater than the chosen threshold.
}
\description{
Check for possible zero-inflation in the data
}
\examples{
# load dasatinib dataset
df <- dasatinib()
# first and last few rows of the dataset
head(df)
tail(df)
# check for zero-inflation with default threshold of 0.5
checkZeroInflation(df)

# change 60\% of the data to zero
df_zero_inflated <- df
df_zero_inflated[1:((nrow(df) * ncol(df) / 3) * 2)] <- 0
checkZeroInflation(df_zero_inflated)

}
