% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{clean_zero_inflation}
\alias{clean_zero_inflation}
\title{Clean zero-inflated genes and cells from the data}
\usage{
clean_zero_inflation(
  data,
  cell_zero_threshold = 0.5,
  gene_zero_threshold = 0.5
)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}
}
\value{
A 'cleaned' version of data, where zero-inflated columns and rows
are removed.
}
\description{
Cleans the data by removing zero-inflated genes (columns) and zero-inflated
cells (rows). A gene or cell is considered zero-inflated if the number of
zero values exceeds the number of zero values expected based on a threshold.
This function ignores NA or NaN values.
}
\examples{
\dontrun{
# load dasatinib dataset
data(dasatinib)
dframe <- dasatinib

# change some of the data to zero
dframe_zero_inflated <- data.frame(dframe)
dframe_zero_inflated[,2:4] <- 0
check_zero_inflation(dframe_zero_inflated)

dframe_cleaned <- clean_zero_inflation(dframe_zero_inflated)
check_zero_inflation(dframe_cleaned)
}

}
