% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{cleanZeroInflation}
\alias{cleanZeroInflation}
\title{Clean zero-inflated genes and cells from the data}
\usage{
cleanZeroInflation(data, cell_zero_threshold, gene_zero_threshold, copy)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}

\item{copy}{A boolean value indicating whether to work on a copy of the data.}
}
\value{
A 'cleaned' version of data. If copy is TRUE, returns a copy of
data but cleaned.
}
\description{
Clean zero-inflated genes and cells from the data
}
\examples{
# load dasatinib dataset
df <- dasatinib()
# first and last few rows of the dataset
head(df)
tail(df)

# change 60\% of the data to zero
df_zero_inflated <- df
df_zero_inflated[1:((nrow(df) * ncol(df) / 3) * 2)] <- 0
checkZeroInflation(df_zero_inflated)

df_cleaned <- cleanZeroInflation(df_zero_inflated, copy = TRUE)
checkZeroInflation(df_cleaned)
print(df_cleaned == df_zero_inflated)

}
