% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{fit_GMM}
\alias{fit_GMM}
\title{Fit a Gaussian Mixture Model to the data}
\usage{
fit_GMM(
  data,
  num_clust = NULL,
  seed = NULL,
  modelNames = NULL,
  inform_progress = TRUE,
  check_zero = TRUE,
  overall_zero_threshold = 0.5,
  cell_zero_threshold = 0.5,
  gene_zero_threshold = 0.5,
  clean_zero = TRUE,
  clean_NA = TRUE
)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell. Data frame passed into this parameter must be
numeric i.e. if a data frame has non-numeric columns, they must be
removed first.}

\item{num_clust}{An integer indicating the numbers of clusters used for GMM
fitting. If NULL, will use the number of features divided by 2. Default is
NULL.}

\item{seed}{An integer value for the seed to be used for random number
generation. Default is NULL, which means that no seed is set and initial
centroids are selected at random.}

\item{modelNames}{A character vector specifying the names of the models to
be used for GMM fitting, as described in \code{?mclust::mclustModelNames}.
Default is NULL, which uses all available models that are suitable for the
data. See \code{?mclust::mclustBIC}, \code{?mclust::mclustICL} or
\code{?mclust::Mclust} for more details.}

\item{inform_progress}{A logical indicating whether to report progress
of GMM fitting while the function is running. Default is TRUE for
interactive sessions, and FALSE otherwise.}

\item{check_zero}{A logical indicating whether to check for zero-inflation.
Default is TRUE.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}

\item{clean_zero}{A logical indicating whether to remove zero-inflated genes.
and cells. Default is TRUE.}

\item{clean_NA}{A logical indicating whether to impute missing values.
Default is TRUE. This can be changed to FALSE if the user wishes to handle
missing values on their own.}
}
\value{
The results of GMM clustering, as an object of class \code{Mclust}. If
multiple models were fitted, the results will provide the optimal model
based on BIC. The components of this object are as described in
\code{?mclust::Mclust}.
}
\description{
Function to conduct Gaussian Mixture Modelling (GMM) onto data. User can
input a seed, otherwise initial centroids are selected at random. This
function can also check the data for missing values or zero-inflation,
and this feature can be turned off if desired. It is recommended to run
\code{clustsimlR::plot_loss()} first, for hyperparameter (\code{num_clust} and
\code{modelNames}) selection. It is also recommended to conduct \strong{PCA} on the
data before GMM, for dimensionality reduction and to reduce the impact
of noise and zero-inflation.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib[, -c(1,2)]  # remove non-numeric and categorical cols

# run GMM clustering on the data
gmm_results <- fit_GMM(df)

# access a summary of the results
summary(gmm_results)

}
\references{
Fraley, C., & Raftery, A. E. (2002). Model-based clustering, discriminant
analysis and density estimation. \emph{Journal of the American Statistical
Association}, 97(458), 611-631.

Fraley, C., & Raftery, A. E. (2007). Bayesian regularization for normal
mixture estimation and model-based clustering. \emph{Journal of
Classification}, 24(2), 155-181.

Jolliffe, I. T. (2002). \emph{Principal Component Analysis}, 2nd Edition,
New York: Springer.
}
