% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_diff.R
\name{intercluster_dist}
\alias{intercluster_dist}
\title{Compute a distance matrix for within-cluster covariances}
\usage{
intercluster_dist(gmm_clustering, plot = TRUE)
}
\arguments{
\item{gmm_clustering}{The output of \code{clustsimlR::fit_GMM}. More
specifically, an object of class \code{Mclust} (see \code{?mclust::Mclust}).}

\item{plot}{A logical value indicating whether to plot a heatmap of the
resulting distance matrix. Default is TRUE.}
}
\value{
A \eqn{\left(G \times G\right)} distance matrix, where \eqn{G} is
the number of clusters. Distances are based on the metric described by
Förstner & Moonen, 1999.
}
\description{
Compute a distance matrix for within-cluster covariances from Gaussian
Mixture Modelling, where each element \eqn{a_{i,j}} is the distance between
the within-cluster covariance matrices of cluster \eqn{i} and \eqn{j}.
Also can plot a heatmap based on this distance matrix. Distances are
computed based on the metric described by Förstner & Moonen, 1999.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib[, -c(1,2)]  # remove non-numeric and categorical cols
df_PCA <- PCA(df, scree_plot = FALSE, pc_heatmap = FALSE)$scores

# run GMM clustering on the data after PCA
gmm_results <- fit_GMM(df_PCA, num_clust = 10)

# compute inter-cluster distances and plot heatmap
intercluster_dist(gmm_results)

}
\references{
Förstner, W., & Moonen, B. (2003). A Metric for Covariance Matrices.
In Geodesy-The Challenge of the 3rd Millennium (pp. 299–309).
Springer Berlin Heidelberg. https://doi.org/10.1007/978-3-662-05296-9_31
}
