% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Conduct Principal Component Analysis}
\usage{
PCA(
  data,
  scree_plot = TRUE,
  pc_heatmap = TRUE,
  check_zero = TRUE,
  overall_zero_threshold = 0.5,
  cell_zero_threshold = 0.5,
  gene_zero_threshold = 0.5,
  clean_zero = TRUE,
  clean_NA = TRUE,
  copy = FALSE
)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell. Data frame passed into this parameter must be
numeric i.e. if a data frame has non-numeric columns, they must be
removed first.}

\item{scree_plot}{A boolean indicating whether to generate a scree plot.
Default is TRUE.}

\item{pc_heatmap}{A boolean indicating whether to generate a heatmap of the
principal components and weights of original features. Default is TRUE.}

\item{check_zero}{A boolean indicating whether to check for zero-inflation.
Default is TRUE.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}

\item{clean_zero}{A boolean indicating whether to remove zero-inflated genes.
and cells. Default is TRUE.}

\item{clean_NA}{A boolean indicating whether to impute missing values.
Default is TRUE. This can be changed to FALSE if the user wishes to handle
missing values on their own.}

\item{copy}{A boolean value indicating whether to work on a copy of the data.
Default is FALSE.}
}
\value{
A list of class \code{princomp} containing the results of the PCA.
The components of the list are as described in \code{?stats::princomp}.
}
\description{
Conduct PCA on the given data. Unless specified otherwise, will also generate
a scree plot of cumulative explained variance and a heatmap comparing the
weights of each variable in the principal component. Also, unless
specified otherwise, will check for missing data and zeroes, and clean them.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib[, -c(1,2)]  # remove non-numeric and categorical cols
pca_results <- PCA(df)

# access first 5 cells and their scores for each PC
pca_results$scores[1:5, ]

}
\references{
Jolliffe, I. T. (2002). \emph{Principal Component Analysis}, 2nd Edition,
New York: Springer.

Upton, G., & Cook, I. (2014). Scree plot. In \emph{A Dictionary of Statistics
(3rd ed.)}. Oxford University Press.
}
