% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{plot_GMM_clusters}
\alias{plot_GMM_clusters}
\title{Plot a t-SNE visualisation for GMM clusters}
\usage{
plot_GMM_clusters(gmm_clustering, seed = NULL, size = 5)
}
\arguments{
\item{gmm_clustering}{An object of class \code{Mclust}, as outputted by
\code{clustsimlR::fit_GMM()}. The components of this object are as described
in \code{?mclust::Mclust}.}

\item{seed}{An integer value for the seed to be used for random number
generation. Default is NULL, which means that no seed is set and initial
embedding points for t-SNE are selected at random. As a result, if not set,
two different t-SNE plots of the same data and same hyperparameters may not
look the same.}

\item{size}{An integer value denoting the size of the points for the t-SNE
plot, in millimeters. Default is 5. For more information see
\code{?ggplot2::size}.}
}
\value{
Returns a t-SNE plot in two dimensions, coloured based on
cluster membership.
}
\description{
Plot a t-SNE visualisation of the data, coloured by GMM clusters.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib[, -c(1,2)]  # remove non-numeric and categorical cols

# run GMM clustering on the data
gmm_results <- fit_GMM(df)

# plot t-SNE
plot_GMM_clusters(gmm_results, seed = 42)

}
\references{
Fraley, C., & Raftery, A. E. (2002). Model-based clustering, discriminant
analysis and density estimation. \emph{Journal of the American Statistical
Association}, 97(458), 611-631.

L.J.P. van der Maaten. (2014). Accelerating t-SNE using tree-based
algorithms. \emph{Journal of Machine Learning Research} 15(Oct), 3221-3245.

van der Maaten, L.J.P. & Hinton, G.E., (2008). Visualizing data using t-SNE.
\emph{Journal of Machine Learning Research}, 9(86), 2579-2605.
}
