% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{plot_loss}
\alias{plot_loss}
\title{Plot the BIC and ICL values for different numbers of clusters and different
models}
\usage{
plot_loss(
  data,
  num_clust = 1:10,
  seed = NULL,
  modelNames = NULL,
  inform_progress = interactive(),
  plot_bic = TRUE,
  plot_icl = TRUE,
  check_zero = TRUE,
  overall_zero_threshold = 0.5,
  cell_zero_threshold = 0.5,
  gene_zero_threshold = 0.5,
  clean_zero = TRUE,
  clean_NA = TRUE
)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell. Data frame passed into this parameter must be
numeric i.e. if a data frame has non-numeric columns, they must be
removed first.}

\item{num_clust}{A vector indicating the numbers of clusters used for the GMM
fitting. Default is 1:10, indicating that GMM will be conducted for 1 to 10
clusters.}

\item{seed}{An integer value for the seed to be used for random number
generation. Default is NULL, which means that no seed is set and initial
centroids are selected at random.}

\item{modelNames}{A character vector specifying the names of the models to
be used for GMM fitting, as described in \code{?mclust::mclustModelNames}.
Default is NULL, which uses all available models that are suitable for the
data. See \code{?mclust::mclustBIC}, \code{?mclust::mclustICL} or
\code{?mclust::Mclust} for more details.}

\item{inform_progress}{A logical indicating whether to report progress
of GMM fitting while the function is running. Default is TRUE for
interactive sessions, and FALSE otherwise.}

\item{plot_bic}{A logical indicating whether to plot BIC values. Default is
TRUE.}

\item{plot_icl}{A logical indicating whether to plot ICL values. Default is
TRUE.}

\item{check_zero}{A logical indicating whether to check for zero-inflation.
Default is TRUE.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}

\item{cell_zero_threshold}{A double representing the maximum proportion of
the data for a single cell (row) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
cell zero-inflated if over half of it are zeroes.}

\item{gene_zero_threshold}{A double representing the maximum proportion of
the data for a single gene (column) that can be zero before being considered
zero-inflated. Default is 0.5, indicating that we consider the data for that
gene zero-inflated if over half of it are zeroes.}

\item{clean_zero}{A logical indicating whether to remove zero-inflated genes.
and cells. Default is TRUE.}

\item{clean_NA}{A logical indicating whether to impute missing values.
Default is TRUE. This can be changed to FALSE if the user wishes to handle
missing values on their own.}
}
\value{
Returns \code{NULL}. Generates BIC and/or ICL plots as specified by user.
}
\description{
Plot the BIC value (Bayesian Information Criterion) the ICL value (Integrated
Complete-data Likelihood) for different numbers of clusters, and different
models used in Gaussian Mixture Modelling (GMM). These values can be used as
a regularised loss function for determining the optimal number of clusters
and optimal model for GMM fitting. User can choose to plot either value,
or both. User can input a seed, otherwise initial centroids are selected at
random.This function can also check the data for missing values or
zero-inflation, and this feature can be turned off if desired. It is also
recommended to conduct PCA on the data before running GMM, for
dimensionality reduction and to reduce the impact of noise and
zero-inflation.
}
\examples{
# load dasatinib dataset
data(dasatinib)
if_data <- dasatinib[ , -c(1,2)]  # remove non-numeric and categorical cols

# plot BIC and ICL values for 1 to 10 clusters
plot_loss(if_data)

}
\references{
Central Limit Theorem. (2008). In The Concise Encyclopedia of Statistics
(pp. 66–68). Springer New York.
https://doi.org/10.1007/978-0-387-32833-1_50

Biernacki, C., Celeux, G., & Govaert, G. (2000). Assessing a mixture model
for clustering with the integrated completed likelihood.
\emph{IEEE Transactions on Pattern Analysis and Machine Intelligence}, 22(7),
719–725. https://doi.org/10.1109/34.865189

Fraley, C., & Raftery, A. E. (2002). Model-based clustering, discriminant
analysis and density estimation. \emph{Journal of the American Statistical
Association}, 97(458), 611-631.

Fraley, C., & Raftery, A. E. (2007). Bayesian regularization for normal
mixture estimation and model-based clustering. \emph{Journal of
Classification}, 24(2), 155-181.

Jolliffe, I. T. (2002). \emph{Principal Component Analysis}, 2nd Edition,
New York: Springer.
}
