% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_PC}
\alias{plot_PC}
\title{Plot data in PC-space}
\usage{
plot_PC(pca_data, pci = 1, pcj = 2)
}
\arguments{
\item{pca_data}{A list of class \code{princomp} containing the results of PCA as
outputted by \code{PCA}. The components of the list are as described in
\code{?stats::princomp}.}

\item{pci}{An integer indicating which principal component to plot as the
x-axis. Default value is 1, i.e. PC1 as the x-axis. If non-integer numeric
values are entered, will be coerced into an integer (rounded down).}

\item{pcj}{An integer indicating which principal component to plot as the
y-axis. Default value is 2, i.e. PC2 as the y-axis. If non-integer numeric
values are entered, will be coerced into an integer (rounded down).}
}
\value{
A ggplot2 scatter plot of of the data in the space defined by
two principal components PCi and PCj.
}
\description{
Plot the scores of the data in the space defined by two principal components
as a scatter plot.
}
\examples{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib[, -c(1,2)]  # remove non-numeric and categorical cols
pca_results <- PCA(df)

# access first 5 cells and their scores for each PC
pca_results$scores[1:5, ]

# Plot data in PC1 vs PC2 space
plot_PC(pca_results)

}
\references{
Jolliffe, I. T. (2002). \emph{Principal Component Analysis}, 2nd Edition,
New York: Springer.
}
