% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clustsimlR.R
\name{run_clustsimlR}
\alias{run_clustsimlR}
\title{Launch Shiny App for clustsimlR}
\usage{
run_clustsimlR()
}
\value{
No return value, but open up a Shiny page.
}
\description{
A function that launches the Shiny app for clustsimlR. Users should note
that the Shiny app is intended for exploratory purposes, and that users
who wish to customise their analyses using this package should use the
functions in this package directly.
}
\examples{
\dontrun{
clustsimlR::run_clustsimlR()
}

}
\references{
Attali, D., Edwards, T. (2024). \emph{shinyalert: Easily Create Pretty Popup
Messages (Modals) in 'Shiny'}. doi:10.32614/CRAN.package.shinyalert
\url{https://doi.org/10.32614/CRAN.package.shinyalert}, R package version
3.1.0, \url{https://CRAN.R-project.org/package=shinyalert}.

Attali, D., Sali, A. (2024). \emph{shinycssloaders: Add Loading Animations to a
'shiny' Output While It's Recalculating}.
doi:10.32614/CRAN.package.shinycssloaders
\url{https://doi.org/10.32614/CRAN.package.shinycssloaders}, R package version
1.1.0, \url{https://CRAN.R-project.org/package=shinycssloaders}.

Chang, W., Cheng, J., Allaire, J., Sievert, C., Schloerke, B., Xie, Y.,
Allen, J., McPherson, J., Dipert, A., Borges, B. (2025). \emph{shiny: Web
Application Framework for R}. doi:10.32614/CRAN.package.shiny
\url{https://doi.org/10.32614/CRAN.package.shiny}, R package version 1.11.1,
\url{https://CRAN.R-project.org/package=shiny}.

Sievert, C., Cheng, J., Aden-Buie, G. (2025). \emph{bslib: Custom 'Bootstrap'
'Sass' Themes for 'shiny' and 'rmarkdown'}. doi:10.32614/CRAN.package.bslib
\url{https://doi.org/10.32614/CRAN.package.bslib}, R package version 0.9.0,
\url{https://CRAN.R-project.org/package=bslib}.
}
