% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{checkZeroInflation}
\alias{checkZeroInflation}
\title{Check for possible zero-inflation in the data}
\usage{
checkZeroInflation(data, overall_zero_threshold = 0.5)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{overall_zero_threshold}{A double representing the maximum proportion of
the data that can be zero before being considered zero-inflated. Default is
0.5, indicating that we consider the data zero-inflated if over half of it
are zeroes.}
}
\value{
A logical value representing whether the proportion of zeroes in the
data is greater than the chosen threshold.
}
\description{
Checks data for possible zero-inflation by comparing the number of zero
values in the data to a threshold (default= 0.5), for which
the data is considered zero-inflated if the number of zero values observed
in the data exceeds the expected number as computed from the threshold. This
function ignores NA or NaN values.
}
\examples{
\dontrun{
# load dasatinib dataset
data(dasatinib)
dframe <- dasatinib

# check for zero-inflation with default threshold of 0.5
checkZeroInflation(dframe)

# change some of the data to zero
dframe_zero_inflated <- data.frame(dframe)
dframe_zero_inflated[, 2:4] <- 0
checkZeroInflation(dframe_zero_inflated)
}

}
