% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checkers.R
\name{cleanMissingData}
\alias{cleanMissingData}
\title{Clean NA or NaN values}
\usage{
cleanMissingData(data, copy = FALSE)
}
\arguments{
\item{data}{A data frame where each column represents a gene and each row
represents a single cell.}

\item{copy}{A boolean indicating whether to work on a copy of the data.}
}
\value{
A 'cleaned' version of the data, with imputed NA and NaN values.
If copy is TRUE, returns a copy of data but cleaned. Default is FALSE.
}
\description{
Cleans the data by imputing NA and NaN values in data using the mean
expression.
}
\examples{
\dontrun{
# load dasatinib dataset
data(dasatinib)
df <- dasatinib

# check for missing values
checkMissingData(df)

# change one of the values to NA
df_one_NA <- data.frame(df)
df_one_NA[1,2] <- NA
df_NA_cleaned <- cleanMissingData(df_one_NA)
checkMissingData(df_NA_cleaned)

#change one of the values to NaN
df_one_NaN <- data.frame(df)
df_one_NaN[1,2] <- 1/0
df_NaN_cleaned <- cleanMissingData(df_one_NaN, copy = TRUE)
checkMissingData(df_NaN_cleaned)
checkMissingData(df_one_NaN)
}
}
